Rename src to lib, build from an index, ignore the build
