+++ index.js
@@ -1,0 +1,7 @@
+module.exports = {
+  apiClient: require('./lib/api-client'),
+  auth: require('./lib/auth'),
+  talkClient: require('./lib/talk-client'),
+  sugar: require('./lib/sugar'),
+  _config: require('./lib/config'),
+};
