+++ lib/config.js
