+++ lib/talk-client.js
